Add a caching MockResponse class for unit testing.
This class is a copy of WrapperAdaptor.GetContentsResponse. That class
was also copied to the SharePoint connector, and very similar classes
written for the file system, Documentum, and OpenText connectors. The
library and SharePoint versions were the most complete; the others
were missing a handful of the flag methods (e.g., setNoFollow).

Only minor changes were made, including some merging of the way the
output stream, metadata, and anchors were handled in the various
implementations. The fields and methods were rearranged to match the
Response source file.
