+++ src/com/google/enterprise/adaptor/WrapperAdaptor.java
@@ -229,187 +229,6 @@ abstract class WrapperAdaptor implements Adaptor {
   }
 
   /**
-   * Counterpart of {@link GetContentsRequest} that allows easy calling of an
-   * {@link Adaptor}. It does not support {@link #respondNotModified}. Be sure
-   * to check {@link #isNotFound()}.
-   */
-  public static class GetContentsResponse implements Response {
-    private OutputStream os;
-    private String contentType;
-    private Date lastModified;
-    private Metadata metadata = new Metadata();
-    private Acl acl;
-    private boolean secure;
-    private List<URI> anchorUris = new ArrayList<URI>();
-    private List<String> anchorTexts = new ArrayList<String>();
-    private boolean notFound;
-    private boolean notModified;
-    private boolean noContent;
-    private boolean noIndex;
-    private boolean noFollow;
-    private boolean noArchive;
-    private URI displayUrl;
-    private boolean crawlOnce;
-    private boolean lock;
-    private Map<String, Acl> fragments = new TreeMap<String, Acl>();
-
-    public GetContentsResponse(OutputStream os) {
-      this.os = os;
-    }
-
-    @Override
-    public void respondNotModified() {
-      notModified = true;
-    }
-
-    @Override
-    public void respondNotFound() {
-      notFound = true;
-    }
-   
-    @Override
-    public void respondNoContent() {
-      noContent = true;
-    }
-
-    @Override
-    public OutputStream getOutputStream() {
-      return os;
-    }
-
-    @Override
-    public void setContentType(String contentType) {
-      this.contentType = contentType;
-    }
-
-    @Override
-    public void setLastModified(Date lastModified) {
-      this.lastModified = lastModified;
-    }
-
-    @Override
-    public void addMetadata(String key, String value) {
-      this.metadata.add(key, value);
-    }
-
-    @Override
-    public void setAcl(Acl acl) {
-      this.acl = acl;
-    }
-
-    @Override
-    public void putNamedResource(String fname, Acl facl) {
-      this.fragments.put(fname, facl);
-    }
-
-    @Override
-    public void setSecure(boolean secure) {
-      this.secure = secure;
-    }
-
-    @Override
-    public void addAnchor(URI uri, String text) {
-      anchorUris.add(uri);
-      anchorTexts.add(text);
-    }
-
-    @Override
-    public void setNoIndex(boolean noIndex) {
-      this.noIndex = noIndex;
-    }
-
-    @Override
-    public void setNoFollow(boolean noFollow) {
-      this.noFollow = noFollow;
-    }
-
-    @Override
-    public void setNoArchive(boolean noArchive) {
-      this.noArchive = noArchive;
-    }
-
-    @Override
-    public void setDisplayUrl(URI displayUrl) {
-      this.displayUrl = displayUrl;
-    }
-
-    @Override
-    public void setCrawlOnce(boolean crawlOnce) {
-      this.crawlOnce = crawlOnce;
-    }
-
-    @Override
-    public void setLock(boolean lock) {
-      this.lock = lock;
-    }
-
-    public String getContentType() {
-      return contentType;
-    }
-
-    public Date getLastModified() {
-      return lastModified;
-    }
-
-    /** Returns reference to modifiable accumulated metadata. */
-    public Metadata getMetadata() {
-      return metadata;
-    }
-
-    public Acl getAcl() {
-      return acl;
-    }
-
-    public boolean isSecure() {
-      return secure;
-    }
-
-    public List<URI> getAnchorUris() {
-      return anchorUris;
-    }
-
-    public List<String> getAnchorTexts() {
-      return anchorTexts;
-    }
-
-    public boolean isNotFound() {
-      return notFound;
-    }
-
-    public boolean isNotModified() {
-      return notModified;
-    }
-    
-    public boolean isNoContent() {
-      return noContent;
-    }
-
-    public boolean isNoIndex() {
-      return noIndex;
-    }
-
-    public boolean isNoFollow() {
-      return noFollow;
-    }
-
-    public boolean isNoArchive() {
-      return noArchive;
-    }
-
-    public URI getDisplayUrl() {
-      return displayUrl;
-    }
-
-    public boolean isCrawlOnce() {
-      return crawlOnce;
-    }
-
-    public boolean isLock() {
-      return lock;
-    }
-  }
-
-  /**
    * Passes through all operations to wrapped {@code DocIdPusher}.
    */
   public static class WrapperDocIdPusher extends AbstractDocIdPusher {
